\name{Sigma-package}
\alias{Sigma-package}
\alias{Sigma}
\docType{package}
\title{A large transition matrix}
\description{
  The A' matrix used in Schoolmaster et al.: Potential Pool-Scale and River-scale Effects of Non-Physical Deterrent Barriers on Invasive Carp in the Illinois River 
}
\details{
This is single function that takes as arguments a set of movment probabilites and a recruitment rate:
(q12, q23, q34, q45, q56, q21, q32, q43, q54, q65, b0). See manuscript for details
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
